Abstract `StatusIcon` into it's own component
