+++ components/StatusUpdates/StatusUpdate.js
@@ -1,47 +1,13 @@
 import PropTypes from "prop-types";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faCheck,
-  faExclamation,
-  faInfo,
-} from "@fortawesome/free-solid-svg-icons";
 
-const iconMapping = {
-  success: {
-    icon: faCheck,
-    style: "bg-green-500 text-white",
-  },
-  warning: {
-    icon: faExclamation,
-    style: "bg-orange-500 text-white",
-  },
-  error: {
-    icon: faExclamation,
-    style: "bg-red-500 text-white",
-  },
-  default: {
-    icon: faInfo,
-    style: "bg-gray-200 text-gray-700",
-  },
-};
+import StatusIcon from "../StatusIcon/StatusIcon";
 
-const StatusUpdate = ({ type, title, description, time }) => {
-  const renderIcon = () => {
-    const style = iconMapping[type].style;
-    const icon = iconMapping[type].icon;
-
-    return (
-      <span
-        className={`flex items-center justify-center h-4 w-4 rounded-full ${style}`}
-      >
-        <FontAwesomeIcon icon={icon} style={{ fontSize: "8px" }} fixedWidth />
-      </span>
-    );
-  };
-
+const StatusUpdate = ({ status, title, description, time }) => {
   return (
     <div className="flex bg-white shadow-sm rounded py-5 px-6">
-      <div className="mr-4 my-0.5">{renderIcon()}</div>
+      <div className="mr-4 my-0.5">
+        <StatusIcon status={status} />
+      </div>
       <div className="">
         <h3 className="c_h-heading mb-1">{title}</h3>
         {description && <p className="text-gray-700">{description}</p>}
@@ -52,7 +18,7 @@ const StatusUpdate = ({ type, title, description, time }) => {
 };
 
 StatusUpdate.propTypes = {
-  type: PropTypes.string.isRequired,
+  status: PropTypes.string.isRequired,
   title: PropTypes.string.isRequired,
   description: PropTypes.string,
   time: PropTypes.string.isRequired,
